Groupcast: Logic methods merged into the cluster class.
