+++ src/app/clusters/groupcast/GroupcastCluster.h
@@ -16,8 +16,16 @@
  */
 #pragma once
 
-#include "GroupcastLogic.h"
+#include "GroupcastContext.h"
+#include <app/AttributeValueEncoder.h>
 #include <app/server-cluster/DefaultServerCluster.h>
+#include <clusters/Groupcast/AttributeIds.h>
+#include <clusters/Groupcast/ClusterId.h>
+#include <clusters/Groupcast/CommandIds.h>
+#include <clusters/Groupcast/Commands.h>
+#include <clusters/Groupcast/Enums.h>
+#include <credentials/GroupDataProvider.h>
+#include <lib/core/CHIPError.h>
 #include <lib/core/DataModelTypes.h>
 #include <protocols/interaction_model/StatusCode.h>
 
@@ -25,19 +33,26 @@ namespace chip {
 namespace app {
 namespace Clusters {
 
+using Status = chip::Protocols::InteractionModel::Status;
+
 /**
  * @brief Provides code-driven implementation for the Groupcast cluster server.
  */
-class GroupcastCluster : public DefaultServerCluster, public GroupcastLogic::Listener
+class GroupcastCluster : public DefaultServerCluster, public Credentials::GroupDataProvider::GroupListener
 {
 public:
-    GroupcastCluster(GroupcastContext && context) :
-        DefaultServerCluster({ kRootEndpointId, Groupcast::Id }), mContext(std::move(context)), mLogic(mContext)
-    {}
-    GroupcastCluster(GroupcastContext && context, BitFlags<Groupcast::Feature> features) :
-        DefaultServerCluster({ kRootEndpointId, Groupcast::Id }), mContext(std::move(context)), mLogic(mContext, features)
-    {}
-    virtual ~GroupcastCluster() {}
+    static constexpr uint16_t kMaxMembershipEndpoints = 255;
+    static constexpr uint16_t kMaxCommandEndpoints    = 20;
+
+    struct EndpointList
+    {
+        EndpointId entries[kMaxMembershipEndpoints];
+        uint16_t count = 0;
+    };
+
+    GroupcastCluster(GroupcastContext && context);
+    GroupcastCluster(GroupcastContext && context, BitFlags<Groupcast::Feature> features);
+    virtual ~GroupcastCluster() override;
 
     CHIP_ERROR Startup(ServerClusterContext & context) override;
     void Shutdown(ClusterShutdownType shutdownType) override;
@@ -51,12 +66,38 @@ public:
                                 ReadOnlyBufferBuilder<DataModel::AcceptedCommandEntry> & builder) override;
 
 private:
-    // GroupcastLogic::Listener implementation
-    void OnMembershipChanged() override;
-    void OnUsedMcastAddrCountChange() override;
+    // GroupDataProvider::GroupListener implementation
+    void OnGroupAdded(FabricIndex fabric_index, const Credentials::GroupDataProvider::GroupInfo & new_group) override;
+    void OnGroupRemoved(FabricIndex fabric_index, const Credentials::GroupDataProvider::GroupInfo & old_group) override;
+
+    // Attribute read methods
+    CHIP_ERROR ReadMembership(const chip::Access::SubjectDescriptor * subject, EndpointId endpoint,
+                              AttributeValueEncoder & aEncoder);
+    CHIP_ERROR ReadMaxMembershipCount(EndpointId endpoint, AttributeValueEncoder & aEncoder);
+    CHIP_ERROR ReadMaxMcastAddrCount(EndpointId endpoint, AttributeValueEncoder & aEncoder);
+    CHIP_ERROR ReadUsedMcastAddrCount(EndpointId endpoint, AttributeValueEncoder & aEncoder);
+    CHIP_ERROR ReadFabricUnderTest(EndpointId endpoint, AttributeValueEncoder & aEncoder);
+
+    // Command handlers
+    Status JoinGroup(FabricIndex fabric_index, const Groupcast::Commands::JoinGroup::DecodableType & data);
+    Status LeaveGroup(FabricIndex fabric_index, const Groupcast::Commands::LeaveGroup::DecodableType & data,
+                      EndpointList & endpoints);
+    Status UpdateGroupKey(FabricIndex fabric_index, const Groupcast::Commands::UpdateGroupKey::DecodableType & data);
+    Status ConfigureAuxiliaryACL(FabricIndex fabric_index, const Groupcast::Commands::ConfigureAuxiliaryACL::DecodableType & data);
+
+    // Helper methods
+    Credentials::GroupDataProvider & Provider() const { return mContext.groupDataProvider; }
+    const chip::FabricTable & Fabrics() const { return mContext.fabricTable; }
+    Status SetKeySet(FabricIndex fabric_index, KeysetId keyset_id, const chip::ByteSpan & key);
+    Status RemoveGroup(FabricIndex fabric_index, GroupId group_id, const Groupcast::Commands::LeaveGroup::DecodableType & data,
+                       EndpointList & endpoints);
+    Status RemoveGroupEndpoint(FabricIndex fabric_index, GroupId group_id, EndpointId endpoint_id, EndpointList & endpoints);
+    uint16_t GetUsedMcastAddrCount() const;
 
     GroupcastContext mContext;
-    GroupcastLogic mLogic;
+    const BitFlags<Groupcast::Feature> mFeatures;
+    DataModel::Provider * mDataModelProvider = nullptr;
+    uint16_t mUsedMcastAddrCount             = 0;
 };
 
 } // namespace Clusters
